reorganize shared/ to unnest dirs
