+++ client/browser/src/libs/code_intelligence/extensions.tsx
@@ -4,9 +4,9 @@ import { combineLatest, from, Observable, Unsubscribable } from 'rxjs'
 import { map, take } from 'rxjs/operators'
 import { ContributableMenu } from '../../../../../shared/src/api/protocol'
 import { TextDocumentDecoration } from '../../../../../shared/src/api/protocol/plainTypes'
-import { CommandListPopoverButton } from '../../../../../shared/src/app/CommandList'
-import { Notifications } from '../../../../../shared/src/app/notifications/Notifications'
 import { Controller as ClientController, createController } from '../../../../../shared/src/client/controller'
+import { CommandListPopoverButton } from '../../../../../shared/src/commandPalette/CommandList'
+import { Notifications } from '../../../../../shared/src/notifications/Notifications'
 import { ConfiguredSubject, SettingsCascade, SettingsCascadeOrError } from '../../../../../shared/src/settings'
 
 import { DOMFunctions } from '@sourcegraph/codeintellify'
