+++ client/browser/src/shared/components/CodeViewToolbar.tsx
@@ -1,8 +1,8 @@
 import H from 'history'
 import * as React from 'react'
 import { Subscription } from 'rxjs'
+import { ActionsNavItems } from '../../../../../shared/src/actions/ActionsNavItems'
 import { ContributableMenu } from '../../../../../shared/src/api/protocol'
-import { ActionsNavItems } from '../../../../../shared/src/app/actions/ActionsNavItems'
 import { ControllerProps } from '../../../../../shared/src/client/controller'
 import { ExtensionsContextProps } from '../../../../../shared/src/context'
 import { ISite, IUser } from '../../../../../shared/src/graphqlschema'
