+++ client/browser/src/shared/components/GlobalDebug.tsx
@@ -1,7 +1,7 @@
 import * as H from 'history'
 import * as React from 'react'
-import { ExtensionStatusPopover } from '../../../../../shared/src/app/ExtensionStatus'
 import { Controller as ClientController } from '../../../../../shared/src/client/controller'
+import { ExtensionStatusPopover } from '../../../../../shared/src/extensions/ExtensionStatus'
 import { sourcegraphUrl } from '../util/context'
 import { ShortcutProvider } from './ShortcutProvider'
