+++ shared/src/actions/ActionItem.tsx
@@ -2,13 +2,13 @@ import H from 'history'
 import * as React from 'react'
 import { from, Subject, Subscription } from 'rxjs'
 import { catchError, map, mapTo, mergeMap, startWith, tap } from 'rxjs/operators'
-import { ExecuteCommandParams } from '../../api/client/providers/command'
-import { ActionContribution } from '../../api/protocol'
-import { urlForOpenPanel } from '../../client/clientCommands'
-import { ControllerProps } from '../../client/controller'
-import { ExtensionsContextProps } from '../../context'
-import { asError, ErrorLike } from '../../errors'
-import { LinkOrButton } from '../../ui/generic/LinkOrButton'
+import { ExecuteCommandParams } from '../api/client/providers/command'
+import { ActionContribution } from '../api/protocol'
+import { urlForOpenPanel } from '../client/clientCommands'
+import { ControllerProps } from '../client/controller'
+import { LinkOrButton } from '../components/LinkOrButton'
+import { ExtensionsContextProps } from '../context'
+import { asError, ErrorLike } from '../errors'
 
 export interface ActionItemProps {
     /**
