+++ shared/src/actions/actions.ts
@@ -1,8 +1,8 @@
 import H from 'history'
-import { TextDocumentItem } from '../../api/client/types/textDocument'
-import { ContributableMenu, Contributions } from '../../api/protocol'
-import { ControllerProps } from '../../client/controller'
-import { ExtensionsContextProps } from '../../context'
+import { TextDocumentItem } from '../api/client/types/textDocument'
+import { ContributableMenu, Contributions } from '../api/protocol'
+import { ControllerProps } from '../client/controller'
+import { ExtensionsContextProps } from '../context'
 
 export interface ActionsProps extends ControllerProps, ExtensionsContextProps {
     menu: ContributableMenu
