--- shared/src/app/app.css
@@ -1,6 +1,0 @@
-@import url('./CommandList.css');
-@import url('./ExtensionStatus.css');
-@import url('./notifications/NotificationItem.css');
-@import url('./notifications/Notifications.css');
-@import url('../ui/generic/PopoverButton.css');
-@import url('../ui/generic/Toggle.css');
