+++ shared/src/client/controller.ts
@@ -6,11 +6,11 @@ import { ExecuteCommandParams } from '../api/client/providers/command'
 import { Contributions, MessageType } from '../api/protocol'
 import { MessageTransports } from '../api/protocol/jsonrpc2/connection'
 import { BrowserConsoleTracer, Trace } from '../api/protocol/jsonrpc2/trace'
-import { ExtensionStatus } from '../app/ExtensionStatus'
-import { Notification } from '../app/notifications/notification'
 import { Context } from '../context'
 import { asError, isErrorLike } from '../errors'
 import { ConfiguredExtension, isExtensionEnabled } from '../extensions/extension'
+import { ExtensionStatus } from '../extensions/ExtensionStatus'
+import { Notification } from '../notifications/notification'
 import { ExtensionManifest } from '../schema/extension.schema'
 import { SettingsCascade } from '../settings'
 import { registerBuiltinClientCommands, updateConfiguration } from './clientCommands'
