+++ shared/src/commandPalette/CommandList.tsx
@@ -6,13 +6,13 @@ import * as React from 'react'
 import { Subscription } from 'rxjs'
 import stringScore from 'string-score'
 import { Key } from 'ts-key-enum'
+import { ActionItem, ActionItemProps } from '../actions/ActionItem'
 import { ContributableMenu, Contributions } from '../api/protocol'
 import { ControllerProps } from '../client/controller'
+import { HighlightedMatches } from '../components/HighlightedMatches'
+import { PopoverButton } from '../components/PopoverButton'
 import { ExtensionsContextProps } from '../context'
-import { HighlightedMatches } from '../ui/generic/HighlightedMatches'
-import { PopoverButton } from '../ui/generic/PopoverButton'
-import { ActionItem, ActionItemProps } from './actions/ActionItem'
-import { getContributedActionItems } from './actions/contributions'
+import { getContributedActionItems } from '../contributions/contributions'
 
 interface Props extends ControllerProps, ExtensionsContextProps {
     /** The menu whose commands to display. */
