+++ shared/src/components/LinkOrSpan.tsx
