+++ shared/src/notifications/NotificationItem.css
