+++ shared/src/shared.css
@@ -1,0 +1,6 @@
+@import url('./commandPalette/CommandList.css');
+@import url('./components/PopoverButton.css');
+@import url('./components/Toggle.css');
+@import url('./extensions/ExtensionStatus.css');
+@import url('./notifications/NotificationItem.css');
+@import url('./notifications/Notifications.css');
