+++ web/src/SourcegraphWebApp.scss
@@ -272,4 +272,4 @@ hr {
 @import './marketing/SurveyPage';
 @import './components/SingleValueCard';
 
-@import './extensions/extensions-client-common';
+@import './extensions/shared';
