+++ web/src/components/DecoratedTextLines.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import VisibilitySensor from 'react-visibility-sensor'
+import { LinkOrSpan } from '../../../shared/src/components/LinkOrSpan'
 import * as GQL from '../../../shared/src/graphqlschema'
-import { LinkOrSpan } from '../../../shared/src/ui/generic/LinkOrSpan'
 import { highlightNode } from '../util/dom'
 
 interface Props {
