+++ web/src/discussions/explore/DiscussionsExploreSection.tsx
@@ -4,10 +4,10 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 import { Observable, Subscription } from 'rxjs'
 import { catchError, map } from 'rxjs/operators'
+import { LinkOrSpan } from '../../../../shared/src/components/LinkOrSpan'
 import { asError, createAggregateError, ErrorLike, isErrorLike } from '../../../../shared/src/errors'
 import { gql } from '../../../../shared/src/graphql'
 import * as GQL from '../../../../shared/src/graphqlschema'
-import { LinkOrSpan } from '../../../../shared/src/ui/generic/LinkOrSpan'
 import { queryGraphQL } from '../../backend/graphql'
 import { ChatIcon } from '../../util/icons'
