+++ web/src/docSite/DocSitePage.tsx
@@ -6,10 +6,10 @@ import { Link, RouteComponentProps } from 'react-router-dom'
 import { Observable, Subject, Subscription } from 'rxjs'
 import { catchError, distinctUntilChanged, map, startWith, switchMap, tap } from 'rxjs/operators'
 import { isError } from 'util'
+import { LinkOrSpan } from '../../../shared/src/components/LinkOrSpan'
 import { asError, createAggregateError, ErrorLike, isErrorLike } from '../../../shared/src/errors'
 import { gql } from '../../../shared/src/graphql'
 import * as GQL from '../../../shared/src/graphqlschema'
-import { LinkOrSpan } from '../../../shared/src/ui/generic/LinkOrSpan'
 import { queryGraphQL } from '../backend/graphql'
 import { HeroPage } from '../components/HeroPage'
 import { Markdown } from '../components/Markdown'
