+++ web/src/enterprise/extensions/registry/RegistryExtensionSourceBadge.tsx
@@ -1,8 +1,8 @@
 import DoNotDisturbIcon from 'mdi-react/DoNotDisturbIcon'
 import WebIcon from 'mdi-react/WebIcon'
 import * as React from 'react'
+import { LinkOrSpan } from '../../../../../shared/src/components/LinkOrSpan'
 import * as GQL from '../../../../../shared/src/graphqlschema'
-import { LinkOrSpan } from '../../../../../shared/src/ui/generic/LinkOrSpan'
 
 export const RegistryExtensionSourceBadge: React.FunctionComponent<{
     extension: Pick<GQL.IRegistryExtension, 'remoteURL' | 'registryName' | 'isLocal'>
