+++ web/src/enterprise/site-admin/dotcom/productSubscriptions/SiteAdminProductSubscriptionNode.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
+import { LinkOrSpan } from '../../../../../../shared/src/components/LinkOrSpan'
 import { gql } from '../../../../../../shared/src/graphql'
 import * as GQL from '../../../../../../shared/src/graphqlschema'
-import { LinkOrSpan } from '../../../../../../shared/src/ui/generic/LinkOrSpan'
 import { AccountName } from '../../../dotcom/productSubscriptions/AccountName'
 import { ProductSubscriptionLabel } from '../../../dotcom/productSubscriptions/ProductSubscriptionLabel'
