+++ web/src/enterprise/user/productSubscriptions/ProductSubscriptionHistory.tsx
@@ -1,8 +1,8 @@
 import format from 'date-fns/format'
 import ExternalLinkIcon from 'mdi-react/ExternalLinkIcon'
 import React from 'react'
+import { LinkOrSpan } from '../../../../../shared/src/components/LinkOrSpan'
 import * as GQL from '../../../../../shared/src/graphqlschema'
-import { LinkOrSpan } from '../../../../../shared/src/ui/generic/LinkOrSpan'
 
 export const ProductSubscriptionHistory: React.FunctionComponent<{
     productSubscription: Pick<GQL.IProductSubscription, 'events'>
