+++ web/src/extensions/ExtensionCard.tsx
@@ -1,12 +1,12 @@
 import WarningIcon from 'mdi-react/WarningIcon'
 import * as React from 'react'
 import { Link } from 'react-router-dom'
+import { LinkOrSpan } from '../../../shared/src/components/LinkOrSpan'
 import { isErrorLike } from '../../../shared/src/errors'
 import { ConfiguredExtension, isExtensionEnabled } from '../../../shared/src/extensions/extension'
-import { SettingsSubject } from '../../../shared/src/graphqlschema'
 import * as GQL from '../../../shared/src/graphqlschema'
+import { SettingsSubject } from '../../../shared/src/graphqlschema'
 import { ExtensionManifest } from '../../../shared/src/schema/extension.schema'
-import { LinkOrSpan } from '../../../shared/src/ui/generic/LinkOrSpan'
 import { ExtensionConfigurationState } from './extension/ExtensionConfigurationState'
 import { WorkInProgressBadge } from './extension/WorkInProgressBadge'
 import { ExtensionsProps, isExtensionAdded, SettingsCascadeProps } from './ExtensionsClientCommonContext'
