+++ web/src/extensions/extension/RegistryExtensionOverviewPage.tsx
@@ -4,8 +4,8 @@ import GithubCircleIcon from 'mdi-react/GithubCircleIcon'
 import * as React from 'react'
 import { RouteComponentProps } from 'react-router'
 import { Link } from 'react-router-dom'
+import { LinkOrSpan } from '../../../../shared/src/components/LinkOrSpan'
 import { isErrorLike } from '../../../../shared/src/errors'
-import { LinkOrSpan } from '../../../../shared/src/ui/generic/LinkOrSpan'
 import { PageTitle } from '../../components/PageTitle'
 import { Timestamp } from '../../components/time/Timestamp'
 import { eventLogger } from '../../tracking/eventLogger'
