+++ web/src/extensions/shared.scss
@@ -1,5 +1,4 @@
-// Imports, styles, and CSS variables for @sourcegraph/extensions-client-common.
-@import '../../shared/src/app/app.css';
+@import '../../shared/src/shared.css';
 
 :root {
     --primary-opacity-2: #{transparentize($primary, 0.8)};
