+++ web/src/repo/GitRef.tsx
@@ -2,10 +2,10 @@ import * as React from 'react'
 import { Link } from 'react-router-dom'
 import { Observable } from 'rxjs'
 import { map } from 'rxjs/operators'
+import { LinkOrSpan } from '../../../shared/src/components/LinkOrSpan'
 import { createAggregateError } from '../../../shared/src/errors'
 import { gql } from '../../../shared/src/graphql'
 import * as GQL from '../../../shared/src/graphqlschema'
-import { LinkOrSpan } from '../../../shared/src/ui/generic/LinkOrSpan'
 import { queryGraphQL } from '../backend/graphql'
 import { Timestamp } from '../components/time/Timestamp'
 import { memoizeObservable } from '../util/memoize'
