+++ web/src/repo/RepoHeader.tsx
@@ -2,8 +2,8 @@ import * as H from 'history'
 import ChevronRightIcon from 'mdi-react/ChevronRightIcon'
 import SettingsIcon from 'mdi-react/SettingsIcon'
 import * as React from 'react'
+import { ActionsNavItems } from '../../../shared/src/actions/ActionsNavItems'
 import { ContributableMenu } from '../../../shared/src/api/protocol'
-import { ActionsNavItems } from '../../../shared/src/app/actions/ActionsNavItems'
 import { ErrorLike, isErrorLike } from '../../../shared/src/errors'
 import * as GQL from '../../../shared/src/graphqlschema'
 import { ActionItem } from '../components/ActionItem'
