+++ web/src/repo/TreePage.tsx
@@ -11,9 +11,9 @@ import * as React from 'react'
 import { Link } from 'react-router-dom'
 import { Observable, Subject, Subscription } from 'rxjs'
 import { catchError, distinctUntilChanged, map, startWith, switchMap, tap } from 'rxjs/operators'
+import { ActionItem } from '../../../shared/src/actions/ActionItem'
+import { ActionsContainer } from '../../../shared/src/actions/ActionsContainer'
 import { ContributableMenu } from '../../../shared/src/api/protocol'
-import { ActionItem } from '../../../shared/src/app/actions/ActionItem'
-import { ActionsContainer } from '../../../shared/src/app/actions/ActionsContainer'
 import { asError, createAggregateError, ErrorLike, isErrorLike } from '../../../shared/src/errors'
 import { gql } from '../../../shared/src/graphql'
 import * as GQL from '../../../shared/src/graphqlschema'
