+++ web/src/repo/blob/LineDecorationAttachment.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import { DecorationAttachmentRenderOptions } from 'sourcegraph'
 import { AbsoluteRepoFile } from '..'
 import { decorationAttachmentStyleForTheme } from '../../../../shared/src/api/client/providers/decoration'
-import { LinkOrSpan } from '../../../../shared/src/ui/generic/LinkOrSpan'
+import { LinkOrSpan } from '../../../../shared/src/components/LinkOrSpan'
 
 interface LineDecorationAttachmentProps extends AbsoluteRepoFile {
     line: number
