+++ web/src/repo/settings/RepoSettingsIndexPage.tsx
@@ -6,10 +6,10 @@ import { RouteComponentProps } from 'react-router'
 import { Link } from 'react-router-dom'
 import { Observable, Subject, Subscription } from 'rxjs'
 import { map, switchMap, tap } from 'rxjs/operators'
+import { LinkOrSpan } from '../../../../shared/src/components/LinkOrSpan'
 import { createAggregateError } from '../../../../shared/src/errors'
 import { gql } from '../../../../shared/src/graphql'
 import * as GQL from '../../../../shared/src/graphqlschema'
-import { LinkOrSpan } from '../../../../shared/src/ui/generic/LinkOrSpan'
 import { queryGraphQL } from '../../backend/graphql'
 import { PageTitle } from '../../components/PageTitle'
 import { Timestamp } from '../../components/time/Timestamp'
